:recycle: #136 Extract catching exceptions to parent classes
